Starting Redux connection
