+++ src/js/redux/actions/search/searchFilterActions.js
@@ -49,6 +49,11 @@ export const updateDomesticForeignSelection = (state) => ({
     selection: state
 });
 
+export const addPOPLocationObject = (state) => ({
+    type: 'ADD_POP_LOCATION_OBJECT',
+    location: state
+});
+
 // Agency Filter
 
 export const updateSelectedAwardingAgencies = (state) => ({
